+++ frontend/src/hooks/auth/useGuestSignIn.ts
@@ -2,10 +2,10 @@ import { useCallback, useContext } from "react";
 import { useNavigate } from "react-router-dom";
 import Cookies from "js-cookie";
 import { AxiosRequestConfig, AxiosResponse } from "axios";
-import { baseAxios } from "../apis/axios";
-import { AuthContext } from "../providers/AuthProvider";
-import { SnackbarContext } from "../providers/SnackbarProvider";
-import { AuthResponse } from "../types/userTypes";
+import { baseAxios } from "../../apis/axios";
+import { AuthContext } from "../../providers/AuthProvider";
+import { SnackbarContext } from "../../providers/SnackbarProvider";
+import { AuthResponse } from "../../types/userTypes";
 
 export const useGuestSignIn = () => {
   const { setIsSignedIn } = useContext(AuthContext);
