+++ frontend/src/hooks/auth/useSignOut.ts
@@ -1,0 +1,51 @@
+import { useContext } from "react";
+import { useNavigate } from "react-router-dom";
+import Cookies from "js-cookie";
+import { AxiosRequestConfig, AxiosResponse } from "axios";
+import { baseAxios } from "../../apis/axios";
+import { AuthContext } from "../../providers/AuthProvider";
+import { SnackbarContext } from "../../providers/SnackbarProvider";
+
+export const useSignOut = () => {
+  const { setIsSignedIn } = useContext(AuthContext);
+  const { handleSetSnackbar } = useContext(SnackbarContext);
+  const navigate = useNavigate();
+
+  const options: AxiosRequestConfig = {
+    url: "/auth/sign_out",
+    method: "DELETE",
+    headers: {
+      "access-token": Cookies.get("_access_token") || "",
+      client: Cookies.get("_client") || "",
+      uid: Cookies.get("_uid") || "",
+    },
+  };
+
+  const handleSignOut = () => {
+    baseAxios(options)
+      .then((res: AxiosResponse) => {
+        console.log(res);
+        Cookies.remove("_access_token");
+        Cookies.remove("_client");
+        Cookies.remove("_uid");
+        setIsSignedIn(false);
+        handleSetSnackbar({
+          open: true,
+          type: "success",
+          message: "ログアウトしました",
+        });
+        navigate("/", { replace: true });
+      })
+      .catch((err) => {
+        console.log(err);
+        handleSetSnackbar({
+          open: true,
+          type: "error",
+          // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
+          message: `${err.response.data.errors[0]}`,
+        });
+      });
+  };
+
+  return handleSignOut;
+};
